+++ R/count_occurrences.R
@@ -8,6 +8,8 @@
 #' counted only once.
 #'
 #' @inheritParams argument_convention
+#' @param drop (`flag`)\cr should non appearing occurrence levels be dropped from the resulting table.
+#'   Note that in that case the remaining occurrence levels in the table are sorted alphabetically.
 #' @param .stats (`character`)\cr statistics to select for the table. Run `get_stats("count_occurrences")`
 #'   to see available statistics for this function.
 #'
@@ -146,7 +148,7 @@ a_count_occurrences <- function(df,
                                 .formats = NULL,
                                 .labels = NULL,
                                 .indent_mods = NULL,
-                                na_str = NA_character_) {
+                                na_str = default_na_str()) {
   denom <- match.arg(denom)
   x_stats <- s_count_occurrences(
     df = df, denom = denom, .N_col = .N_col, .df_row = .df_row, drop = drop, .var = .var, id = id
@@ -209,10 +211,12 @@ a_count_occurrences <- function(df,
 #' @order 2
 count_occurrences <- function(lyt,
                               vars,
+                              id = "USUBJID",
+                              drop = TRUE,
                               var_labels = vars,
                               show_labels = "hidden",
                               riskdiff = FALSE,
-                              na_str = NA_character_,
+                              na_str = default_na_str(),
                               nested = TRUE,
                               ...,
                               table_names = vars,
@@ -225,15 +229,16 @@ count_occurrences <- function(lyt,
   extra_args <- list(
     .stats = .stats, .formats = .formats, .labels = .labels, .indent_mods = .indent_mods, na_str = na_str
   )
+  s_args <- list(id = id, drop = drop, ...)
 
   if (isFALSE(riskdiff)) {
-    extra_args <- c(extra_args, list(...))
+    extra_args <- c(extra_args, s_args)
   } else {
     extra_args <- c(
       extra_args,
       list(
         afun = list("s_count_occurrences" = a_count_occurrences),
-        s_args = list(...)
+        s_args = s_args
       )
     )
   }
@@ -274,8 +279,10 @@ count_occurrences <- function(lyt,
 #' @order 3
 summarize_occurrences <- function(lyt,
                                   var,
+                                  id = "USUBJID",
+                                  drop = TRUE,
                                   riskdiff = FALSE,
-                                  na_str = NA_character_,
+                                  na_str = default_na_str(),
                                   ...,
                                   .stats = "count_fraction_fixed_dp",
                                   .formats = NULL,
@@ -286,15 +293,16 @@ summarize_occurrences <- function(lyt,
   extra_args <- list(
     .stats = .stats, .formats = .formats, .labels = .labels, .indent_mods = .indent_mods, na_str = na_str
   )
+  s_args <- list(id = id, drop = drop, ...)
 
   if (isFALSE(riskdiff)) {
-    extra_args <- c(extra_args, list(...))
+    extra_args <- c(extra_args, s_args)
   } else {
     extra_args <- c(
       extra_args,
       list(
         afun = list("s_count_occurrences" = a_count_occurrences),
-        s_args = list(...)
+        s_args = s_args
       )
     )
   }