+++ R/response_biomarkers_subgroups.R
@@ -68,6 +68,7 @@
 #' @name response_biomarkers_subgroups
 tabulate_rsp_biomarkers <- function(df,
                                     vars = c("n_tot", "n_rsp", "prop", "or", "ci", "pval"),
+                                    na_str = default_na_str(),
                                     .indent_mods = 0L) {
   checkmate::assert_data_frame(df)
   checkmate::assert_character(df$biomarker)
@@ -79,6 +80,7 @@ tabulate_rsp_biomarkers <- function(df,
     tab_sub <- h_tab_rsp_one_biomarker(
       df = df_sub,
       vars = vars,
+      na_str = na_str,
       .indent_mods = .indent_mods
     )
     # Insert label row as first row in table.