+++ tests/testthat/test-survival_duration_subgroups.R
@@ -188,7 +188,9 @@ testthat::test_that("tabulate_survival_subgroups works correctly with both `n_to
   testthat::expect_snapshot(res)
 
   # Check header of table.
-  result_header_both_survtime <- to_string_matrix(result_both_survtime)[2, ]
+  result_header_both_survtime <- to_string_matrix(result_both_survtime,
+    with_spaces = FALSE, print_txt_to_copy = FALSE
+  )[2, ]
 
   res <- testthat::expect_silent(result_header_both_survtime)
   testthat::expect_snapshot(res)
@@ -203,6 +205,26 @@ testthat::test_that("d_survival_subgroups_colvars functions as expected with val
     method = "p-value (log-rank)",
     time_unit = "Months"
   )
+
+  res <- testthat::expect_silent(result)
+  testthat::expect_snapshot(res)
+})
+
+testthat::test_that("tabulate_survival_subgroups na_str argument works as expected", {
+  adtte <- adtte_local
+
+  df <- extract_survival_subgroups(
+    variables = list(tte = "AVAL", is_event = "is_event", arm = "ARM", subgroups = c("SEX", "BMRKR2")),
+    data = adtte
+  )
+  df$hr$hr[2:5] <- NA
+
+  result <- basic_table() %>%
+    tabulate_survival_subgroups(
+      df,
+      time_unit = adtte$AVALU[1],
+      na_str = "<No data>"
+    )
 
   res <- testthat::expect_silent(result)
   testthat::expect_snapshot(res)