+++ NEWS.md
@@ -1,4 +1,18 @@
-# tern 0.9.2.9015
+# tern 0.9.3.9007
+
+### New Features
+* Refactored `g_forest` to output a `ggplot` object instead of a `grob` object.
+* Added `h_glm_negbin` to `h_glm_count` to enable count data analysis using a negative binomial model.
+* Added argument `grade_groups_only` to `count_occurrences_by_grade` to allow users to only display rows for specified grade groups.
+
+### Bug Fixes
+* Fixed nested column split label overlay issue in `rtable2gg` to clean up appearance of text labels.
+* Fixed bug in `s_ancova` causing incorrect difference calculations for arm variables with irregular levels.
+
+### Miscellaneous
+* Added function `expect_snapshot_ggplot` to test setup file to process plot snapshot tests and allow plot dimensions to be set.
+
+# tern 0.9.3
 
 ### New Features
 * Added `ref_group_position` function to place the reference group facet last, first or at a certain position.
@@ -13,18 +27,25 @@
 * Added referential footnotes to `surv_time` for censored range observations, controlled via the `ref_fn_censor` parameter.
 * Added helper function `h_adlb_abnormal_by_worst_grade` to prepare `ADLB` data to use as input in `count_abnormal_by_worst_grade`.
 * Added `s_bland_altman` function to assess agreement between two numerical vectors. 
+* Added function `rtable2gg` that converts `rtable` objects to `ggplot` objects.
+* Added helper function to set default `na_str` globally with `set_default_na_str()` and added `default_na_str()` for all interested functions.
 
 ### Enhancements
 * Added `ref_group_coxph` parameter to `g_km` to specify the reference group used for pairwise Cox-PH calculations when `annot_coxph = TRUE`.
 * Added `annot_coxph_ref_lbls` parameter to `g_km` to enable printing the reference group in table labels when `annot_coxph = TRUE`.
 * Added `x_lab` parameter to `g_lineplot` to customize x-axis label.
 * Remove 25% padding of y-axis in `g_lineplot`.
+* Added support for creating multiple risk difference columns, each comparing to a single comparison group. Multiple comparison groups can be specified as a vector via the `arm_y` argument.
+* Allowed numeric vector as `count_by` input in `analyze_num_patients` and `summarize_num_patients`.
+* Aligned plot and table vertically in `g_lineplot`.
 
 ### Bug Fixes
 * Fixed bug in `decorate_grob` preventing text wrapping from accounting for font size.
+* Fixed implementation of `na_str` argument in all column-wise analysis and tabulation functions.
 
 ### Miscellaneous
 * Specified minimal version of package dependencies.
+* Upgraded `to_string_matrix` to take into account `widths` and other printing parameters.
 
 # tern 0.9.2